examples/requirements.txt
# A2A Agent Examples Dependencies

# HTTP Client
httpx>=0.27.0
requests>=2.31.0

# Web3 Interaction
web3>=6.15.0
eth-account>=0.10.0

# Async Support
asyncio>=3.4.3
aiohttp>=3.9.0

# CLI and Formatting
rich>=13.7.0          # Beautiful terminal output
click>=8.1.7          # CLI interface
colorama>=0.4.6       # Terminal colors

# Data Handling
pydantic>=2.5.0       # Data validation
python-dotenv>=1.0.0  # Environment variables

# Database
motor>=3.3.0          # Async MongoDB driver

# Utilities
tenacity>=8.2.3       # Retry logic
loguru>=0.7.2         # Enhanced logging

